requirements.txt
astroid==2.2.5
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mypy==0.740
mypy-extensions==0.4.1
pylint==2.3.1
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4
wrapt==1.11.2
pillow==6.2.0
pygame==2.0.0.dev4
gevent==1.4.0
dataclasses==0.7
